Change OM type to String in device-renderer.yang
JIRA: TRNSPRTPCE-885

+++ renderer/src/main/java/org/opendaylight/transportpce/renderer/openconfiginterface/OpenConfigInterface190.java
@@ -34,9 +34,10 @@ import org.opendaylight.yang.gen.v1.http.openconfig.net.yang.transport.line.comm
 import org.opendaylight.yang.gen.v1.http.openconfig.net.yang.transport.line.common.rev190603.transport.line.common.port.top.optical.port.ConfigBuilder;
 import org.opendaylight.yang.gen.v1.http.openconfig.net.yang.transport.types.rev210729.AdminStateType;
 import org.opendaylight.yang.gen.v1.http.openconfig.net.yang.transport.types.rev210729.FrequencyType;
-import org.opendaylight.yang.gen.v1.http.org.opendaylight.transportpce.device.renderer.rev250325.ServicePathInput;
+import org.opendaylight.yang.gen.v1.http.org.opendaylight.transportpce.device.renderer.rev251001.ServicePathInput;
 import org.opendaylight.yang.gen.v1.http.org.opendaylight.transportpce.portmapping.rev250902.mapping.Mapping;
 import org.opendaylight.yang.gen.v1.urn.ietf.params.xml.ns.yang.iana._if.type.rev230126.EthernetCsmacd;
+import org.opendaylight.yangtools.yang.common.Uint16;
 
 public class OpenConfigInterface190 {
 
@@ -242,7 +243,7 @@ public class OpenConfigInterface190 {
                 .yang.gen.v1.http.openconfig.net.yang.terminal.device.rev210729.terminal.optical.channel.top.optical
                 .channel.ConfigBuilder();
         configBuilder.setTargetOutputPower(input.getTargetOutputPower());
-        configBuilder.setOperationalMode(input.getOperationalMode());
+        configBuilder.setOperationalMode(Uint16.valueOf(input.getOperationalMode()));
         if (input.getCenterFreq() != null) { //input.getCenterFreq().getValue().intValue()
             configBuilder.setFrequency(FrequencyType
                     .getDefaultInstance(String.valueOf((long) (input.getCenterFreq().getValue()
